Restructure (P)Config to follow new paradigm
