+++ src/Core/PConfig/Type/JitPConfig.php
@@ -19,52 +19,53 @@
  *
  */
 
-namespace Friendica\Core\PConfig;
+namespace Friendica\Core\PConfig\Type;
 
-use Friendica\Core\BasePConfig;
+use Friendica\Core\PConfig\Cache\Cache;
 use Friendica\Model;
 
 /**
- * This class implements the preload configuration, which will cache
- * all user config values per call in a cache.
+ * This class implements the Just-In-Time configuration, which will cache
+ * user config values in a cache, once they are retrieved.
  *
- * Minimizes the number of database queries to retrieve configuration values at the cost of memory.
+ * Default Configuration type.
+ * Provides the best performance for pages loading few configuration variables.
  */
-class PreloadPConfig extends BasePConfig
+class JitPConfig extends BasePConfig
 {
-	/** @var array */
-	private $config_loaded;
+	/**
+	 * @var array Array of already loaded db values (even if there was no value)
+	 */
+	private $db_loaded;
 
 	/**
-	 * @param Cache                $configCache The configuration cache
-	 * @param Model\Config\PConfig $configModel The configuration model
+	 * @param Cache                                 $configCache The configuration cache
+	 * @param \Friendica\Core\PConfig\Model\PConfig $configModel The configuration model
 	 */
-	public function __construct(Cache $configCache, Model\Config\PConfig $configModel)
+	public function __construct(Cache $configCache, \Friendica\Core\PConfig\Model\PConfig $configModel)
 	{
 		parent::__construct($configCache, $configModel);
-		$this->config_loaded = [];
+		$this->db_loaded = [];
 	}
 
 	/**
 	 * {@inheritDoc}
 	 *
-	 * This loads all config values everytime load is called
-	 *
 	 */
 	public function load(int $uid, string $cat = 'config')
 	{
-		// Don't load the whole configuration twice or with invalid uid
-		if (!$uid || !empty($this->config_loaded[$uid])) {
+		// If not connected or no uid, do nothing
+		if (!$uid || !$this->configModel->isConnected()) {
 			return;
 		}
 
-		// If not connected, do nothing
-		if (!$this->configModel->isConnected()) {
-			return;
-		}
+		$config = $this->configModel->load($uid, $cat);
 
-		$config                    = $this->configModel->load($uid);
-		$this->config_loaded[$uid] = true;
+		if (!empty($config[$cat])) {
+			foreach ($config[$cat] as $key => $value) {
+				$this->db_loaded[$uid][$cat][$key] = true;
+			}
+		}
 
 		// load the whole category out of the DB into the cache
 		$this->configCache->load($uid, $config);
@@ -81,15 +82,19 @@ class PreloadPConfig extends BasePConfig
 			return $default_value;
 		}
 
-		if (empty($this->config_loaded[$uid])) {
-			$this->load($uid);
-		} elseif ($refresh) {
-			if ($this->configModel->isConnected()) {
-				$config = $this->configModel->get($uid, $cat, $key);
-				if (isset($config)) {
-					$this->configCache->set($uid, $cat, $key, $config);
-				}
+		// if the value isn't loaded or refresh is needed, load it to the cache
+		if ($this->configModel->isConnected() &&
+		    (empty($this->db_loaded[$uid][$cat][$key]) ||
+		     $refresh)) {
+
+			$dbvalue = $this->configModel->get($uid, $cat, $key);
+
+			if (isset($dbvalue)) {
+				$this->configCache->set($uid, $cat, $key, $dbvalue);
+				unset($dbvalue);
 			}
+
+			$this->db_loaded[$uid][$cat][$key] = true;
 		}
 
 		// use the config cache for return
@@ -107,10 +112,6 @@ class PreloadPConfig extends BasePConfig
 			return false;
 		}
 
-		if (empty($this->config_loaded[$uid])) {
-			$this->load($uid);
-		}
-
 		// set the cache first
 		$cached = $this->configCache->set($uid, $cat, $key, $value);
 
@@ -120,6 +121,8 @@ class PreloadPConfig extends BasePConfig
 		}
 
 		$stored = $this->configModel->set($uid, $cat, $key, $value);
+
+		$this->db_loaded[$uid][$cat][$key] = $stored;
 
 		return $cached && $stored;
 	}
@@ -133,11 +136,11 @@ class PreloadPConfig extends BasePConfig
 			return false;
 		}
 
-		if (empty($this->config_loaded[$uid])) {
-			$this->load($uid);
-		}
-
 		$cacheRemoved = $this->configCache->delete($uid, $cat, $key);
+
+		if (isset($this->db_loaded[$uid][$cat][$key])) {
+			unset($this->db_loaded[$uid][$cat][$key]);
+		}
 
 		if (!$this->configModel->isConnected()) {
 			return $cacheRemoved;
