+++ tests/src/Core/Lock/DatabaseLockDriverTest.php
@@ -22,11 +22,10 @@
 namespace Friendica\Test\src\Core\Lock;
 
 use Friendica\Core\Lock\DatabaseLock;
-use Friendica\Factory\ConfigFactory;
+use Friendica\Core\Config\Factory\ConfigFactory;
 use Friendica\Test\DatabaseTestTrait;
 use Friendica\Test\Util\Database\StaticDatabase;
 use Friendica\Test\Util\VFSTrait;
-use Friendica\Util\ConfigFileLoader;
 use Friendica\Util\Profiler;
 use Mockery;
 use Psr\Log\NullLogger;
