+++ tests/src/Core/StorageManagerTest.php
@@ -23,22 +23,22 @@ namespace Friendica\Test\src\Core;
 
 use Dice\Dice;
 use Friendica\Core\Config\IConfig;
-use Friendica\Core\Config\PreloadConfig;
+use Friendica\Core\Config\Type\PreloadConfig;
 use Friendica\Core\Hook;
 use Friendica\Core\L10n;
 use Friendica\Core\Session\ISession;
 use Friendica\Core\StorageManager;
 use Friendica\Database\Database;
 use Friendica\DI;
-use Friendica\Factory\ConfigFactory;
-use Friendica\Model\Config\Config;
+use Friendica\Core\Config\Factory\ConfigFactory;
+use Friendica\Core\Config\Model\Config;
 use Friendica\Model\Storage;
 use Friendica\Core\Session;
 use Friendica\Network\HTTPClient;
 use Friendica\Test\DatabaseTest;
 use Friendica\Test\Util\Database\StaticDatabase;
 use Friendica\Test\Util\VFSTrait;
-use Friendica\Util\ConfigFileLoader;
+use Friendica\Core\Config\Cache\ConfigFileLoader;
 use Friendica\Util\Profiler;
 use org\bovigo\vfs\vfsStream;
 use Psr\Log\LoggerInterface;
